# src/swaggerUI.json
{
    "swagger": "2.0",
    "info": {
      "description": "XMeme Operational APIs",
      "version": "1.0.0",
      "title": "CWoD XMEME API",
      "contact": {
        "email": "arjun.garg.cse01@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "virtserver.swaggerhub.com",
    "basePath": "/XMEME7/XMEME/1.0.0",
    "tags": [
      {
        "name": "XMEME",
        "description": "Operations available to regular developers"
      }
    ],
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
      "/memes": {
        "get": {
          "tags": [
            "XMEME"
          ],
          "summary": "Fetch latest 100 memes",
          "description": "Fetch the latest 100 memes created from collection",
          "operationId": "GetAllMemes",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "409": {
              "description": "Duplicate POST requests with the same payload (name/url/caption) "
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "post": {
          "tags": [
            "XMEME"
          ],
          "summary": "Add a new meme to collection",
          "description": "Add a new meme to collection",
          "operationId": "addInventory",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Meme object that needs to be added in the collection",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Meme"
              }
            }
          ],
          "responses": {
            "409": {
              "description": "Duplicate POST requests with the same payload (name/url/caption) "
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/memes/:id": {
        "get": {
          "tags": [
            "XMEME"
          ],
          "summary": "Get a meme from collection",
          "description": "Get a meme corresponding to a particular Id",
          "operationId": "GetMemeByID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Meme id that needs to be fetched from the collection",
              "required": true,
              "schema": {
                "$ref": "#/definitions/"
              }
            }
          ],
          "responses": {
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "tags": [
            "XMEME"
          ],
          "summary": "Delete a meme from collection",
          "description": "Delete a meme from collection corresponding to a particular Id",
          "parameters": [],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "patch": {
          "tags": [
            "XMEME"
          ],
          "summary": "Update an existing meme in collection",
          "description": "Update an existing meme in collection",
          "operationId": "updateMeme",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Meme object that needs to be updated in the collection",
              "required": true,
              "schema": {
                "$ref": "#/definitions/"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "definitions": {
      "Meme": {
        "type": "object",
        "required": [
          "caption",
          "name",
          "url"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Arjun Garg"
          },
          "caption": {
            "type": "string",
            "example": "This is a meme"
          },
          "url": {
            "type": "string",
            "example": "https://i.redd.it/uyrdtpig1o831.jpg"
          }
        }
      }
    }
  }